+++ changelog.md
@@ -4,6 +4,8 @@
 
 - [#0001](https://github.com/ignite/cli/pull/0001) Add nodetime
 - [#0003](https://github.com/ignite/ignite-files/pull/3) Add Hermes binaries
+- [#0012](https://github.com/ignite/ignite-files/pull/12) Update `nodetime` deps and add log level support to `ts-relayer`
+
 
 ### Features